FinanceAgent/pyproject.toml
[project]
name = "financeagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "flask>=3.1.3",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
]
